pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigic-geonode-importer"
version = "1.1.0.post1"
description = "GeoNode Importer Django app"
requires-python = ">=3.11"
dependencies = [
    "Django>=4.2,<5.0",
    "djangorestframework==3.14.0",
    "celery==5.4.0",
    "django-filter==24.2",
    "geonode-django-dynamic-model==0.4.0",
    "pandas"
]

[project.urls]
Repository = "https://github.com/CentroGeo/sigic-geonode-importer"